Random access into xorshift128+
Following Vigna's https://arxiv.org/abs/1404.0390, a fixed size jump in
xorshift128+ can be computed by precomputed a GF(2) polynomial modulo
the characteristic polynomial of one step of xorshift128+ (which is
linear as a transform on GF(2)^128).

We can jump by a variable amount by precomputing these polynomials for
each power of two jump, then multiplying them all together.
Unfortunately, multiplication of polynomials modulo the characteristic
polynomial is slow, since that modulus is dense.  Fortunately, all
finite fields of a given order are isomorphic, so we can do all the
multiplications in the equivalent field modulo the pentanomial
x^128 + x^7 + x^2 + x + 1.  This is reasonably fast if we take advantage
of the carryless multiply instructions on recent CPUs.  We then convert
back to the slow modulus using a precomputed linear map corresponding to
the finite field isomorphism.

If the base point is fixed, we can compose the GF(2^128) -> GF(2^128)
linear map with the linear map corresponding to the final jump, for some
extra speedup.

Computing these linear maps requires finding an isomorphism from the
fast finite field to the slow one.  gen_xorshift does this using an
algorithm of Allombert.

The result is that we can use xorshift128+ as a reasonably efficient
counter mode generator: seeking into the appropriate part of the stream
and using sequential mode from there.

Caveat: This code is extremely nonportable.  If it's ever to actually be
used, slower versions for less capable hardware will have to be written,
ideally using runtime detection of which instructions exist.